Move newtype wrappers to hschain-db
